+++ app/Plugin/Setup.php
@@ -24,6 +24,13 @@ use ConnectorForPropstack\Propstack\PostTypes\ImmoObject;
  */
 class Setup {
 	/**
+	 * Collect the errors.
+	 *
+	 * @var array<int,WP_Error>
+	 */
+	private array $import_errors = array();
+
+	/**
 	 * Instance of this object.
 	 *
 	 * @var ?Setup
@@ -460,6 +467,9 @@ class Setup {
 				break;
 			}
 		} while ( $import_obj->has_load_more() );
+
+		// remember the errors for the completion text.
+		$this->import_errors = $import_obj->get_errors();
 	}
 
 	/**
@@ -479,7 +489,15 @@ class Setup {
 		update_option( 'esfw_step', absint( get_option( 'esfw_max_steps' ) ) );
 
 		// prepare the completed text.
-		$completed_text = '<strong>' . __( 'Setup has been run.', 'connector-for-propstack' ) . '</strong> ' . __( 'Your objects from Propstack has been imported. Click on "Completed" to view them.', 'connector-for-propstack' );
+		if ( ! empty( $this->import_errors ) ) {
+			$completed_text = '<strong>' . __( 'Setup has been run, but the import reported problems.', 'connector-for-propstack' ) . '</strong> ';
+
+			foreach ( $this->import_errors as $error ) {
+				$completed_text .= '<br>' . $error->get_error_message();
+			}
+		} else {
+			$completed_text = '<strong>' . __( 'Setup has been run.', 'connector-for-propstack' ) . '</strong> ' . __( 'Your objects from Propstack has been imported. Click on "Completed" to view them.', 'connector-for-propstack' );
+		}
 
 		/**
 		 * Filter the text for display if the setup has been run.